Refactored solutions led to same & improved performances

+++ submissions/CountSubarraysWithCostLessThanOrEqualToK.java
@@ -2,7 +2,7 @@
 
 class CountSubarraysWithCostLessThanOrEqualToK {
     public long countSubarrays(int[] nums, long k) {
-        TreeMap<Long, Integer> window = new TreeMap<>();
+        TreeMap<Integer, Integer> window = new TreeMap<>();
         boolean isTest = false;
         int left = 0;
         long result = 0;
@@ -12,10 +12,10 @@ class CountSubarraysWithCostLessThanOrEqualToK {
         }
         for (int right = 0; right < nums.length; right++) {
             long length = right - left + 1;
-            window.merge((long) nums[right], 1, Integer::sum);
+            window.merge(nums[right], 1, Integer::sum);
 
             while ((window.lastKey() - window.firstKey()) * length > k) {
-                window.computeIfPresent((long) nums[left++], (key, value) -> value == 1 ? null : value - 1);
+                window.computeIfPresent(nums[left++], (key, value) -> value == 1 ? null : value - 1);
                 length--;
             }
             if (isTest) {
